Added new proto field definitions using Wrappers to accomodate null's and missing vals.  Cleaned up python.  Started to make test for json -> proto conversion.
