+++ .gitignore
@@ -42,6 +42,7 @@ yarn-error.log
 plots/out**
 
 *__pycache__
-*_pb2.py
 *.pyc
+proto/*_pb2.*
+proto/*_pb.*
 proto/*.json
