+++ package.json
@@ -21,7 +21,7 @@
     "build-all:task:windows": "yarn build-extension && yarn build-viewer",
     "build-extension": "cd ./lighthouse-extension && yarn build",
     "build-viewer": "cd ./lighthouse-viewer && yarn build",
-    "clean": "rimraf proto/*.json proto/*_pb2.py proto/*_pb2.pyc proto/__pycache__ *.report.html *.report.dom.html *.report.json *.devtoolslog.json *.trace.json || true",
+    "clean": "rimraf proto/*.json proto/*_pb2.* proto/*_pb.* proto/__pycache__ proto/*.pyc *.report.html *.report.dom.html *.report.json *.devtoolslog.json *.trace.json || true",
     "lint": "[ \"$CI\" = true ] && eslint --quiet -f codeframe . || eslint .",
     "smoke": "node lighthouse-cli/test/smokehouse/run-smoke.js",
     "debug": "node --inspect-brk ./lighthouse-cli/index.js",
@@ -64,7 +64,8 @@
     "ultradumbBenchmark": "./lighthouse-core/scripts/benchmark.js",
     "mixed-content": "./lighthouse-cli/index.js --chrome-flags='--headless' --preset=mixed-content",
     "minify-latest-run": "./lighthouse-core/scripts/lantern/minify-trace.js ./latest-run/defaultPass.trace.json ./latest-run/defaultPass.trace.min.json && ./lighthouse-core/scripts/lantern/minify-devtoolslog.js ./latest-run/defaultPass.devtoolslog.json ./latest-run/defaultPass.devtoolslog.min.json",
-    "compile-proto": "protoc --python_out=./ ./proto/lighthouse-result.proto && python ./proto/proto_lhr_validator.py"
+    "compile-proto": "protoc --python_out=./ ./proto/lighthouse-result.proto",
+    "build-proto": "cd proto/ && export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=cpp && PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION_VERSION=2 && python make_python_roundtrip.py"
   },
   "devDependencies": {
     "@types/chrome": "^0.0.60",
